+++ locale/en.json
@@ -33,7 +33,6 @@
 	"other_section_title": "Your learning plan",
 	"home_page_suggestions": "Tailored suggestions are based on your grade and other profile preferences:",
 	"home_page_learning_plan_link": "See more learning in your Learning plan",
-	"home_page_suggestions_link": "See more suggestions for you",
 	"required_learning_done": "You have completed all your required learning.",
 	"no_planned_learning_yet": "You don't have any planned learning yet.",
 	"learning_plan_page_title": "Learning plan",
@@ -47,7 +46,7 @@
 	"item_control_book": "Book now",
 	"item_control_cancel": "Cancel booking",
 	"item_control_add": "Add to learning plan",
-	"suggestions_page_title": "Suggestions for you",
+	"course_catalogue_page_title": "Course catalogue",
 	"suggestion_section_title": "From name_of_selected_work_area",
 	"learning_record_page_title": "Learning record",
 	"learning_page_feedback_intro": "You can leave feedback on X item(s).",
