+++ src/lib/config/index.ts
@@ -159,6 +159,7 @@ export const HOMEPAGE_COMPLETE_LEARNING_PLAN_COURSES = env.HOMEPAGE_COMPLETE_LEA
 
 export const SUGGESTIONS_MAX_COURSES = env.SUGGESTIONS_MAX_COURSES || null
 export const SUGGESTIONS_EXCLUDE_LEARNING_PLAN_COURSES = env.SUGGESTIONS_EXCLUDE_LEARNING_PLAN_COURSES || false
+export const COURSE_CATALOGUE_A_Z_MAX_COURSES = Number(env.COURSE_CATALOGUE_A_Z_MAX_COURSES) || 20
 
 export const GRADE_PRIORITY_ORDER_ENABLED = env.GRADE_PRIORITY_ORDER_ENABLED === 'true'
 export const GRADE_PRIORITY_ORDER = env.GRADE_PRIORITY_ORDER.split(',') || 'AA,AO,EO,HEO,SEO,G7,G6,PB1,PB2,PB3,PS'
