+++ src/lib/service/cslService/models/learning/learningPlan/basicCourseResponse.ts
@@ -1,0 +1,11 @@
+import {Type} from 'class-transformer'
+import {BasicCourse} from './basicCourse'
+
+export class BasicCourseResponse {
+	@Type(() => BasicCourse)
+	results: BasicCourse[]
+
+	page: number
+	totalResults: number
+	size: number
+}
