+++ src/lib/service/cslService/models/learning/suggestions/profileSuggestionsResponse.ts
@@ -1,5 +1,5 @@
 import {Type} from 'class-transformer'
-import {SuggestionSection} from '../../../../../../ui/controllers/suggestions/model/suggestionSection'
+import {SuggestionSection} from '../../../../../../ui/controllers/courseCatalogue/model/suggestionSection'
 
 export class ProfileSuggestionsResponse {
 	@Type(() => SuggestionSection)
