+++ src/lib/utils/ui/response.ts
@@ -1,0 +1,6 @@
+export interface Response<T> {
+	results: T[]
+	page: number
+	totalResults: number
+	size: number
+}
