+++ src/lib/utils/ui/searchParams.ts
@@ -1,0 +1,3 @@
+export interface SearchParams {
+	getAsUrlParams(page?: number): string | undefined
+}
