+++ src/ui/controllers/courseCatalogue/model/profilePreferencesPage.ts
