+++ src/ui/controllers/courseCatalogue/model/suggestionSection.ts
