+++ views/assets/styles/common/_global.scss
@@ -39,6 +39,14 @@ input:focus, textarea:focus, select:focus, button:focus, .button:focus {
 	@include grid-column(1/6);
 }
 
+.column-one-tenth {
+	@include grid-column(1/10);
+}
+
+.column-nine-tenths {
+	@include grid-column(9/10);
+}
+
 .search-lede {
 	font-size: 19px;
 }
