+++ views/nunjucks/components/courseCard/actions/macro.njk
@@ -40,7 +40,7 @@
         Add to learning plan <span class="visually-hidden">: {{ params.courseTitle }}</span>
     {% endset %}
     {{ ctaLink({
-        "href": "/suggestions-for-you/add/" + params.courseId + "?ref=" + params.ref,
+        "href": "/course-catalogue/profile-preferences/add/" + params.courseId + "?ref=" + params.ref,
         "html": html
     }) }}
 {% endmacro %}
