+++ views/nunjucks/components/learningCard/header/macro.njk
@@ -16,6 +16,7 @@
             "text": params.courseTitle
         },
         "moduleCount": params.moduleCount,
-        "description": params.shortDescription
+        "description": params.shortDescription,
+        "headingTag": params.headingTag
     }) }}
 {% endmacro %}
