+++ views/nunjucks/course-catalogue/profile-preferences/index.njk
@@ -1,0 +1,20 @@
+{% extends '../layout.njk' %}
+{% import "../components/suggestionsCourseCard/macro.njk" as suggestionsCard %}
+
+{% set anchor = 'profile-preferences' %}
+{% block catalogueContent %}
+    <div class='grid-row'>
+        <div class='column-full'>
+            {% for section in sections %}
+                <h2 class='heading heading-medium {% if loop.index0 == 0 %}heading--no-margin-top{% endif %}'>From {{ section.title }}</h2>
+                {% if section.courses | length > 0 %}
+                    {% for course in section.courses %}
+                        {{ suggestionsCard.profileSuggestionsCard(course) }}
+                    {% endfor %}
+                {% else %}
+                    <p class='govuk-body'>There are no suggested courses for this section at this time.</p>
+                {% endif %}
+            {% endfor %}
+        </div>
+    </div>
+{% endblock %}
