+++ views/nunjucks/course/layout.njk
@@ -13,7 +13,7 @@
             {% if pageModel.backLink %}
                 {{ backLink(pageModel.backLink, "Back") }}
             {% else %}
-                {{ backLink("/suggestions-for-you", "Find another course") }}
+                {{ backLink("/course-catalogue", "Find another course") }}
             {% endif %}
         </p>
         {% if pageModel.status === "Archived" %}
