+++ views/nunjucks/home/index.njk
@@ -54,7 +54,7 @@
                     </div>
                 {% else %}
                     <p class="govuk-body">There is currently no learning in your plan.</p>
-                    <p class="govuk-body">You can add learning by checking <a href="/suggestions-for-you">Suggestions for you</a> or <a href="/search?q=">searching for a specific course</a>.</p>
+                    <p class="govuk-body">You can add learning by checking <a href="/course-catalogue">the course catalogue</a> or <a href="/search?q=">searching for a specific course</a>.</p>
                 {% endif %}
             </div>
         </div>
