+++ views/nunjucks/root/nav.njk
@@ -4,7 +4,7 @@
     <ul id="proposition-links">
         {% if signedInUser.hasCompleteProfile() %}
             {{ navbarLink(originalUrl, "/home", "Home") }}
-            {{ navbarLink(originalUrl, "/suggestions-for-you", i18n('suggestions_page_title')) }}
+            {{ navbarLink(originalUrl, "/course-catalogue", i18n('course_catalogue_page_title')) }}
             {{ navbarLink(originalUrl, "/learning-record", "Learning record") }}
             {{ navbarLink(originalUrl, "/profile", "Profile") }}
             {% if signedInUser.isAdmin() %}
