+++ views/nunjucks/search/index.njk
@@ -105,7 +105,7 @@
                         </ul>
                         <h2 class="heading-medium">You can:</h2>
                         <ul class="no-results">
-                            <li>view <a href="/suggestions-for-you">suggestions for you</a></li>
+                            <li>view <a href="/course-catalogue">the course catalogue</a></li>
                         </ul>
                     {% endif %}
                 </div>
