--- views/nunjucks/suggestions-for-you/components/suggestionsCourseCard/macro.njk
@@ -1,39 +1,0 @@
-{% import components + "/learningCard/actions/macro.njk" as learningCardActions %}
-{% import components + "/learningCard/details/macro.njk" as learningCardDetails %}
-{% import components + "/learningCard/header/macro.njk" as learningCardHeader %}
-{% import components + "/learningCard/content/macro.njk" as learningCardContent %}
-
-{% macro suggestionsCard(params) %}
-    <div class="learning-card">
-        <div class='learning-card__header'>
-            {{ learningCardHeader.courseCardHeader({
-                "type": params.type,
-                "courseId": params.id,
-                "courseTitle": params.title
-            }) }}
-            {{ learningCardDetails.basicDetails({
-                type: params.type,
-                duration: params.duration,
-                cost: params.costInPounds
-            }) }}
-        </div>
-        {{ learningCardContent.learningCardContent({
-            "moduleCount": params.moduleCount,
-            "shortDescription": params.shortDescription
-        }) }}
-        {% set primary %}
-            {% if params.status == 'IN_PROGRESS' %}
-                {{ learningCardActions.alreadyInLearningPlan() }}
-            {% else %}
-                {{ learningCardActions.addToLearningPlan({
-                    "courseId": params.id,
-                    "courseTitle": params.title,
-                    "ref": "suggestions"
-                }) }}
-            {% endif %}
-        {% endset %}
-        {{ learningCardActions.learningCardActions({
-            "primary": primary
-        }) }}
-    </div>
-{% endmacro %}
