--- views/nunjucks/suggestions-for-you/index.njk
@@ -1,48 +1,0 @@
-{% extends baseLayout %}
-{% from components + "/search/macro.njk" import searchHeader %}
-{% from components + "/banner/macro.njk" import renderBanners %}
-{% from "./components/suggestionsCourseCard/macro.njk" import suggestionsCard %}
-
-{% set pageTitle = i18n('suggestions_page_title') %}
-
-{% block content %}
-    <div class="container container--main">
-        <div class="grid-row ">
-            <div class="column-full no-border">
-                {% set header %}
-                    <h1 class="heading heading-large heading--flush">
-                        {{ pageTitle }}
-                    </h1>
-                {% endset %}
-                {{ searchHeader({
-                    "headerHtml": header
-                }) }}
-                <p class='govuk-body'>
-                    Your suggestions are more accurate when you have an updated profile.
-                </p>
-                <div class="change-settings-container">
-                    <p class="govuk-body">You can change the preferences in <a href="/profile">your profile page</a>.</p>
-                </div>
-                <h2 class='heading heading-medium'>Browse courses</h2>
-                <div class='margin-bottom-7'>
-                    <a class='section-link section-link--active' href='#'>Profile preferences</a>
-                </div>
-                {{ renderBanners(banners) }}
-                <div class='grid-row'>
-                    <div class='column-full'>
-                        {% for section in sections %}
-                            <h2 class='heading heading-medium {% if loop.index0 == 0 %}heading--no-margin-top{% endif %}'>From {{ section.title }}</h2>
-                            {% if section.courses | length > 0 %}
-                                {% for course in section.courses %}
-                                    {{ suggestionsCard(course) }}
-                                {% endfor %}
-                             {% else %}
-                                <p class='govuk-body'>There are no suggested courses for this section at this time.</p>
-                            {% endif %}
-                        {% endfor %}
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
