+++ views/page/booking/choose-date.html
@@ -100,7 +100,7 @@
                 {/if}
 
                 {#if !course.record}
-                    <a class="link--block" href="/suggestions-for-you/add/{course.id}?ref=home">
+                    <a class="link--block" href="/course-catalogue/profile-preferences/add/{course.id}?ref=home">
                         Add <span class="visuallyhidden">{course.title} </span>to learning plan
                     </a>
                 {/if}
