+++ views/page/booking/payment-options.html
@@ -26,7 +26,7 @@
                             <BookingCancellationPolicy />
                             <input type="submit" class="button" value="Continue">
                             <input type="hidden" name="_csrf" value="{_csrf}">
-                            <a class="link--block" href="/suggestions-for-you/add/{course.id}?ref=home">
+                            <a class="link--block" href="/course-catalogue/profile-preferences/add/{course.id}?ref=home">
                                 Add <span class="visually-hidden">{course.title}</span> to learning plan
                             </a>
                         </div>
